requirements.txt
joblib==1.3.2
pandas==1.5.3
numpy==1.23.2
yt-dlp==2023.7.6
tqdm==4.64.1
jsonschema==4.16.0
requests==2.28.2
bs4==0.0.1
fake-useragent==1.4.0
aria==0.0.1b0
etter_profanity==0.7.0